fix: resolvendo problema de inicialização dos cômodos na home_page.dart

+++ lib/models/sala_comodo.dart
@@ -2,7 +2,7 @@ import 'package:appsmarthome/models/comodo.dart';
 
 class SalaModel extends Comodo {
   String nome;
-  int luminosidade;
+  double luminosidade;
   bool presencaDetectada;
 
   SalaModel({
